use unverified SSLContext with --skip-cert-verify
Using the skip-cert-verify flag sets the verify attribute of the request
library function calls to true which conflicts with the default context
requested from the SSL library that was introduced in #158. In order to
circumvent this issue an unverified SSLContext is used on a conditional
basis depending on whether the skip-cert-verify flag is defined.